+++ tests/lib.rs
@@ -2,8 +2,8 @@ use rawcmd::{Context, Command, Intent, Result};
 
 #[test]
 fn performs_command() {
-    fn resolver0(_: &Intent, _: &mut Context) -> Result<usize> { Ok(0) };
-    fn resolver1(_: &Intent, _: &mut Context) -> Result<usize> { Ok(1) };
+    fn resolver0(_: &Intent, _: &mut Context) -> Result<i32> { Ok(0) };
+    fn resolver1(_: &Intent, _: &mut Context) -> Result<i32> { Ok(1) };
     let result = Command::with_name("0")
         .with_subcommand(
             Command::with_name("1").with_resolver(resolver0)
